Update patch set 1
Patch Set 1: Code-Review-1

please try to add the following file to your device tree overlay:

overlay/frameworks/base/core/res/res/values-mmc310-mnc120/config.xml

with the following in it:

    <integer name="config_volte_replacement_rat">6</integer>

This should fix the situation where voice radio tech reports as LTE but Voice Over LTE (VoLTE) is not supported (Sprint network). By setting it to 6 (1x) it defaults the voice tech to CDMAPhone and avoids the being set to GSMPhone causing voice calls to fail. 

This also avoids the issues created by setting rilV7NeedCDMALTEPhone. Which breaks Global and GSM support for your device. This way the quirk is only used when a Sprint SIM is present in the device. Allowing phone object to be updated normally with other operators SIMs in the device.

Patch-set: 1