Update patch set 2
Patch Set 2:

The LTE tile didn't do anything up until recently. The code that got merged got some back and forth discussion about the best way of figuring out which network modes to use on the various possible radio combos, and it's possible it's wrong for this model. Are you having issues switching to LTE, from LTE, or both?

Patch-set: 2