Update patch set 2
Patch Set 2:

LTE was on, tried toggling it off. It worked along with completely killing my connection. I tried everything to get it back from playing with different dialer codes etc but what I ended up doing was complete wipe, clean flash CM and restore efs backup. If you see the old phone overlay being deleted, it had something in place for the LTE tile but I checked and that doesn't seem to be used anymore. I was thinking of a situation where on my sprint note 2, the same thing would happen and sbrissen fixed it by adding this as a prop "ro.config.qc_lte_network_modes=true" but that doesn't seem to be used right now either

Patch-set: 2