Update patch set 2
Patch Set 2:

It builds, but we're still testing it. (I don't have an LS980 in front of me)

https://github.com/lg-devs/android_kernel_lge_msm8974/tree/cm-12.0/include/linux/nfc

Yes the kernel exports it.

The HAL originates from other devices that use the pn544 chip. As for it working as is on klte, I'm not sure. I know I still need to change a few things for it on LGE.

Patch-set: 2